Permutation (nPr)
nPr Calculation using Recursion
This C++ program calculates the permutation (nPr) for two given integers n and r. The formula for permutation is:

𝑛
𝑃
𝑟
=
𝑛
!
(
𝑛
−
𝑟
)
!
nPr= 
(n−r)!
n!
​
 
Where:

n! represents the factorial of n.

r is the number of objects selected from a total of n objects.

nPr is the number of possible ways to arrange r objects from a set of n distinct objects.

How It Works
The program uses a recursive function fact() to calculate the factorial of a number.

It takes the values n and r as input from the user.

The formula for permutation (nPr) is then applied using the factorial function.

The result is displayed to the user.