True-False Permutations
Generate All Combinations of True/False
This C++ program generates all possible combinations of true and false values for a given number of elements. It recursively explores all the possibilities and prints each combination.

How It Works
The program takes an integer n as input, which represents the number of elements.

It initializes an array com[] of size n to store the boolean values (true or false).

The function generateTF() uses recursion to fill the array with all possible combinations of true and false for n elements.

Each time the array is filled, it prints the current combination.

The process continues until all combinations are generated.

Code Details
The function generateTF() takes three parameters:

com[]: An array to store the current combination of true and false.

n: The number of elements.

index: The current position in the array.

If index reaches n, the function prints the current combination.

The function then recursively sets the current index to false, calls itself to fill the next index, then sets the index to true and calls itself again.

Usage
Compile and run the program.

Input the number of elements (n) when prompted.

The program will output all possible combinations of true and false values for n elements.