Factorial of a number
This C++ program calculates the factorial of a number using a recursive function. A factorial of a number n is the product of all positive integers less than or equal to n. It is represented as n!.

In this program, the function factorial(int n) is implemented using recursion. It takes an integer n and returns the factorial of that number.

Functionality:
Factorial Calculation: The factorial function recursively calls itself to compute the factorial of a given number.

Base Case: If n is 0 or 1, the factorial is 1 (since 0! = 1 and 1! = 1).

Recursive Case: For values greater than 1, it multiplies the number n with the factorial of (n-1) until the base case is reached.

The main() function initializes the number n = 8, calls the factorial() function, and prints the result.